Standardise type to size_t from unsigned int

+++ Cache.cpp
@@ -1,55 +1,55 @@
 #include "Cache.hpp"
-Cache::Cache(unsigned int cacheSize, unsigned int associativity, unsigned int blockSize){
+Cache::Cache(size_t cacheSize, size_t associativity, size_t blockSize){
     this->associativity = associativity;
     this->blockSize = blockSize;
     this->cache = std::vector<CacheSet>(cacheSize/(associativity*blockSize), CacheSet(associativity));
     this->cacheSize = cacheSize/(associativity*blockSize); //TODO: rename this variable
 }
 
-unsigned int Cache::getIndex(unsigned int address) {
+size_t Cache::getIndex(size_t address) {
     // this returns the index of the memory
     return (address/blockSize) % cacheSize;
 }
 
-unsigned int Cache::getTag(unsigned int address) {
+size_t Cache::getTag(size_t address) {
     return (address/blockSize) / cacheSize;
 }
 
-bool Cache::checkCacheLine(unsigned int address){
+bool Cache::checkCacheLine(size_t address){
     // get the respective address
     // loop inside the cacheset list to find cacheline
-    unsigned int index = getIndex(address);
-    unsigned int tag = getTag(address);
+    size_t index = getIndex(address);
+    size_t tag = getTag(address);
     return cache[index].checkCacheLine(tag);
 }
 
-void Cache::invalidateCacheLine(unsigned int address){
+void Cache::invalidateCacheLine(size_t address){
     // performs a checkCacheLine
     // performs a invalidateCacheLine
-    unsigned int index = getIndex(address);
-    unsigned int tag = getTag(address);
+    size_t index = getIndex(address);
+    size_t tag = getTag(address);
     cache[index].invalidateCacheLine(tag);
 }
 
 
-void Cache::addCacheLine(unsigned int address, unsigned int state){
+void Cache::addCacheLine(size_t address, size_t state){
     // performs a addCacheLine
-    unsigned int index = getIndex(address);
-    unsigned int tag = getTag(address);
+    size_t index = getIndex(address);
+    size_t tag = getTag(address);
     cache[index].addCacheLine(tag,state);
 }
 
-bool Cache::readCacheLine(unsigned int address){
+bool Cache::readCacheLine(size_t address){
     // checkCacheLine
     // readCacheLine
-    unsigned int index = getIndex(address);
-    unsigned int tag = getTag(address);
+    size_t index = getIndex(address);
+    size_t tag = getTag(address);
     return cache[index].readCacheLine(tag);
 }
 
-bool Cache::updateCacheLine(unsigned int address,unsigned int state){
+bool Cache::updateCacheLine(size_t address,size_t state){
     // checkCacheLine
-    unsigned int index = getIndex(address);
-    unsigned int tag = getTag(address);
+    size_t index = getIndex(address);
+    size_t tag = getTag(address);
     return cache[index].updateCacheLine(tag,state);
 }
